Infer values across notebook cells (#1968)
## Changes
When linting python code, infer values using not only code from current
cell but also code from previous cells

### Linked issues
Progresses #1912
Progresses #1205 

### Functionality 
None

### Tests
- [x] manually tested
- [x] added unit tests

Resolved 60 out of 891 "cannot be computed" advices when running make
solacc

---------
