+++ tests/unit/source_code/samples/values_across_cells.py
@@ -1,0 +1,6 @@
+# Databricks notebook source
+a = 12
+
+# COMMAND ----------
+
+spark.table(f"{a}")
