+++ tests/unit/source_code/test_notebook_linter.py
@@ -1,3 +1,5 @@
+from pathlib import Path
+
 import pytest
 from databricks.sdk.service.workspace import Language
 
@@ -546,4 +548,22 @@ def test_notebook_linter_tracks_use(extended_test_index, lang, source, expected)
     linter = NotebookLinter.from_source(extended_test_index, source, lang)
     assert linter is not None
     advices = list(linter.lint())
+    assert advices == expected
+
+
+def test_computes_values_across_cells(extended_test_index, mock_path_lookup):
+    path = mock_path_lookup.resolve(Path("values_across_cells.py"))
+    source = path.read_text()
+    linter = NotebookLinter.from_source(extended_test_index, source, Language.PYTHON)
+    advices = list(linter.lint())
+    expected = [
+        Advice(
+            code='table-migrate',
+            message='The default format changed in Databricks Runtime 8.0, from Parquet to Delta',
+            start_line=5,
+            start_col=0,
+            end_line=5,
+            end_col=19,
+        )
+    ]
     assert advices == expected
